+++ openai-audio.el
@@ -82,7 +82,8 @@ TEMPERATURE, and LANGUAGE."
   "Send translate audio request.
 
 Argument FILE is the audio file to translate, in one of these formats: mp3, mp4,
-mpeg, mpga, m4a, wav, or webm. CALLBACK is the execuation after request is made.
+mpeg, mpga, m4a, wav, or webm.  CALLBACK is the execuation after request is
+made.
 
 Arguments CONTENT-TYPE, KEY and ORG-ID are global options; however, you
 can overwrite the value by passing it in.